#993. IntPtrPointer tests added

+++ LibTest/ffi/Pointer/IntPtrPointer_A02_t01.dart
@@ -3,13 +3,21 @@
 // BSD-style license that can be found in the LICENSE file.
 
 /// @assertion void operator []=(int index, int value)
-/// The 16-bit two's complement integer at address + 2 * index.
 ///
-/// A Dart integer is truncated to 16 bits (as if by .toSigned(16)) before being
-/// stored, and the 16-bit value is sign-extended when it is loaded.
+/// The 32 or 64-bit two's complement integer at address + (4 or 8) * index.
+///
+/// On 32-bit platforms this is a 32-bit integer, and on 64-bit platforms this
+/// is a 64-bit integer.
+///
+/// On 32-bit platforms a Dart integer is truncated to 32 bits (as if by
+/// .toSigned(32)) before being stored, and the 32-bit value is sign-extended
+/// when it is loaded.
+///
+/// On 32-bit platforms the address must be 4-byte aligned, and on 64-bit
+/// platforms the address must be 8-byte aligned.
 ///
 /// @description Check that operator []=(int index, int value) sets int at
-/// address + 2 * index.
+/// address + (4 or 8) * index.
 /// @author sgrekhov@unipro.ru
 
 import "dart:ffi";
@@ -17,9 +25,9 @@ import "package:ffi/ffi.dart";
 import '../../../Utils/expect.dart';
 
 void main() {
-  Pointer<Int16> p1 = calloc<Int16>(2);
+  Pointer<IntPtr> p1 = calloc<IntPtr>(2);
   try {
-    Pointer<Int16> p2 = new Pointer.fromAddress(p1.address + sizeOf<Int16>());
+    Pointer<IntPtr> p2 = new Pointer.fromAddress(p1.address + sizeOf<IntPtr>());
     p1[0] = 1;
     p1[1] = 42;
     Expect.equals(1, p1[0]);
